requirements.txt
streamlit
google-generativeai
SpeechRecognition
pydub
moviepy
